second normalize BankAccount.institution -> Institution

+++ src/modules/bank-module/resolvers/bank/delete_bank_account/delete_bank_account.resolver.ts
@@ -1,10 +1,20 @@
-import { Resolver, Mutation, Arg, UseMiddleware } from 'type-graphql';
-import { ApiResponse } from 'neutronpay-wallet-common/dist/shared';
+import {
+	Resolver,
+	Mutation,
+	Arg,
+	UseMiddleware,
+	Ctx,
+} from 'type-graphql';
+import {
+	ApiResponse,
+	CustomMessage,
+} from 'neutronpay-wallet-common/dist/shared';
 import { InjectRepository } from 'typeorm-typedi-extensions';
-import { BankAccountRepository } from '../../../repository/BankAccountRepository';
+import { BankAccountRepository } from '../../../repository';
 import { DeleteBankAccountDto } from './delete_bank_account.dto';
 import { BankAccount } from '../../../entity';
 import { isAuth } from 'neutronpay-wallet-common/dist/middleware';
+import { BankGQLContext } from '../../../server';
 
 export const ApiDeleteBankAccount = ApiResponse<String>(
 	'DeleteBankAccount',
@@ -22,9 +32,27 @@ class DeleteBankAccountResolver {
 	@UseMiddleware(isAuth)
 	@Mutation(() => ApiDeleteBankAccount, { nullable: true })
 	async deleteBankAccount(
+		@Ctx() { currentUser }: BankGQLContext,
 		@Arg('data') { bankAccountId }: DeleteBankAccountDto,
 	): Promise<ApiDeleteBankAccountType> {
 		try {
+			const bankAccount = this.bankAccountRepository.findOne({
+				where: {
+					id: bankAccountId,
+					userId: currentUser?.userId,
+				},
+			});
+			if (!bankAccount) {
+				return {
+					success: false,
+					errors: [
+						{
+							path: 'currentUser',
+							message: CustomMessage.stopBeingNaughty,
+						},
+					],
+				};
+			}
 			await this.bankAccountRepository.delete({
 				id: bankAccountId,
 			});
